mlc: add rule that disables unused drivers
run line (from minimize_linux_config):
./run_on_vm.sh $TARGET $RULE/target.sh | while read module; do $RULE/mod2conf.sh "$module"; done | python3 run_pylkc.py --pylkc=$PYLKC --rule=$RULE:run"

Where
RULE: the rule directory (rule04_driver)
TARGET: the user and ip of the vm (root@30.0.0.2)
PYLKC: path from which to import pylkc library (~/Dev/pylkc/python3)

The target.sh script feeds the emits unused driver modules, which are then
fed to mod2conf through the for loop.

The mod2conf script looks for the CONFIG_* option for the module it's given as
a command line argument by searching through the kernel makefiles for a line
that adds said module to obj-$(CONFIG_*).

The pylkc rule is the same rule for rules 02 and 03, which disables options
given to it through stdin

Signed-off-by: Tal Shorer <tal.shorer@gmail.com>